backend/requirements.txt
duckdb
tabulate
openai==0.28
flask
flask_cors
python-dotenv
flask-mysqldb
jwt
cryptography
flask_mail
minio